Make the source for GCC4 build and run with VS2013
Make the source for GCC4 build and run with VS2013

Now it works with Visual Studio Community 2013 Update 5. Tested on
Windows 8.1 Pro x64.

The 3rd party libs are available at
https://www.dropbox.com/s/ohfdxp8qa1ba36e/3rdParty_v4.0-VS2013.zip?dl=0
