--- Source/GCC4/Msvc/GameCode4_2010.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "resource.h"
