+++ Source/GCC4/Msvc/GameCode4_2013.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "resource.h"
