+++ Source/GCC4/Msvc/GameCode4_2013.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.40629.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GameCode4_2013", "GameCode4_2013.vcxproj", "{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Debug|x64 = Debug|x64
+		Release|Win32 = Release|Win32
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Win32.ActiveCfg = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Win32.Build.0 = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|x64.ActiveCfg = Debug|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|x64.Build.0 = Debug|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Win32.ActiveCfg = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Win32.Build.0 = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|x64.ActiveCfg = Release|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
