+++ Source/TeapotWars/Msvc/TeapotWars_2013.sln
@@ -1,0 +1,94 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.40629.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TeapotWars_2013", "TeapotWars_2013.vcxproj", "{B44F8785-5D50-4326-9929-43CC020068FB}"
+	ProjectSection(ProjectDependencies) = postProject
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD} = {5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GameCode4_2013", "..\..\GCC4\Msvc\GameCode4_2013.vcxproj", "{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GCC4EditorDLL_2013", "..\..\GCC4Editor\Msvc\GCC4EditorDLL_2013.vcxproj", "{01B47F4F-C2A9-4C16-8C8B-B4846428089A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD} = {5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeapotWarsEditorApp_2013", "..\..\GCC4Editor\EditorApp\TeapotWarsEditorApp_2013.csproj", "{6183905E-44A0-4419-9288-B78B96CC1841}"
+	ProjectSection(ProjectDependencies) = postProject
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A} = {01B47F4F-C2A9-4C16-8C8B-B4846428089A}
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD} = {5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|Win32 = Debug|Win32
+		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|Win32 = Release|Win32
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|Mixed Platforms.Build.0 = Debug|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|Win32.ActiveCfg = Debug|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|Win32.Build.0 = Debug|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|x64.ActiveCfg = Debug|x64
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Debug|x64.Build.0 = Debug|x64
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|Any CPU.ActiveCfg = Release|x64
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|Mixed Platforms.ActiveCfg = Release|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|Mixed Platforms.Build.0 = Release|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|Win32.ActiveCfg = Release|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|Win32.Build.0 = Release|Win32
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|x64.ActiveCfg = Release|x64
+		{B44F8785-5D50-4326-9929-43CC020068FB}.Release|x64.Build.0 = Release|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Mixed Platforms.Build.0 = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Win32.ActiveCfg = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|Win32.Build.0 = Debug|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|x64.ActiveCfg = Debug|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Debug|x64.Build.0 = Debug|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Any CPU.ActiveCfg = Release|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Mixed Platforms.ActiveCfg = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Mixed Platforms.Build.0 = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Win32.ActiveCfg = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|Win32.Build.0 = Release|Win32
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|x64.ActiveCfg = Release|x64
+		{5F8DE669-F90C-498F-891B-EE6ACE0CA1BD}.Release|x64.Build.0 = Release|x64
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|Any CPU.ActiveCfg = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|Mixed Platforms.Build.0 = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|Win32.ActiveCfg = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|Win32.Build.0 = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Debug|x64.ActiveCfg = Debug|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|Any CPU.ActiveCfg = Release|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|Mixed Platforms.ActiveCfg = Release|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|Mixed Platforms.Build.0 = Release|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|Win32.ActiveCfg = Release|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|Win32.Build.0 = Release|Win32
+		{01B47F4F-C2A9-4C16-8C8B-B4846428089A}.Release|x64.ActiveCfg = Release|Win32
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|Win32.Build.0 = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Win32.ActiveCfg = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|Win32.Build.0 = Release|Any CPU
+		{6183905E-44A0-4419-9288-B78B96CC1841}.Release|x64.ActiveCfg = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
